# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/teacher": {
      "post": {
        "operationId": "TeacherController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Teacher"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Teacher"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TeacherController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Teacher"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Teacher"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/teacher/{id}": {
      "get": {
        "operationId": "TeacherController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Teacher"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Teacher"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "TeacherController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Teacher"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Teacher"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TeacherController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Teacher"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "School Compass 365",
    "description": "swagger json: <a href=\"/swagger.json\">swagger.json</a>",
    "version": "1.0",
    "contact": {},
    "license": {
      "name": "Mozilla Public License Version 2.0",
      "url": "https://github.com/vaibhavdubay/School-Compass-365-be/blob/main/LICENSE"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "enum": [
              "super_admin",
              "admin",
              "teacher",
              "student"
            ],
            "type": "string"
          },
          "changePassword": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "userName",
          "password",
          "role",
          "changePassword",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "admin",
              "teacher",
              "student"
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "userName",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8
          },
          "role": {
            "enum": [
              "super_admin",
              "admin",
              "teacher",
              "student"
            ],
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "userName",
          "password",
          "role"
        ]
      },
...
      "AcademicYear": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "academicYear": {
            "type": "string"
          },
          "current": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "academicYear",
          "current",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "filename": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "admin",
              "teacher",
              "student"
            ]
          },
          "mimeType": {
            "type": "string"
          },
          "buffer": {
            "type": "object"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "role",
          "mimeType",
          "buffer",
          "userId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Class": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "nextClass": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "streamsRequired": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "className",
          "nextClass",
          "order",
          "streamsRequired",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "School": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "establishedYear": {
            "type": "number"
          },
          "address1": {
            "type": "string"
          },
          "address2": {
            "type": "string"
          },
          "academicYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicYear"
            }
          },
          "logo": {
            "$ref": "#/components/schemas/Image"
          },
          "logoUrl": {
            "type": "string"
          },
          "currentAcademicYear": {
            "$ref": "#/components/schemas/AcademicYear"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pincode": {
            "type": "number"
          },
          "schoolDISECode": {
            "type": "string"
          },
          "schoolCode": {
            "type": "string"
          },
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phoneNumber",
          "establishedYear",
          "address1",
          "address2",
          "academicYears",
          "logo",
          "logoUrl",
          "currentAcademicYear",
          "city",
          "state",
          "pincode",
          "schoolDISECode",
          "schoolCode",
          "classes",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "Teacher@12"
          },
          "image": {
            "type": "string",
            "format": "binary"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "years_of_experience": {
            "type": "number"
          }
        },
        "required": [
          "password",
          "firstName",
          "lastName",
          "email",
          "phoneNumber",
          "userName",
          "subjects",
          "years_of_experience"
        ]
      },
      "Teacher": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other"
            ],
            "type": "string"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "academicYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicYear"
            }
          },
          "profileImage": {
            "$ref": "#/components/schemas/Image"
          },
          "profileImageUrl": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "years_of_experience": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "gender",
          "school",
          "firstName",
          "lastName",
          "email",
          "phoneNumber",
          "academicYears",
          "profileImage",
          "profileImageUrl",
          "user",
          "subjects",
          "years_of_experience",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "Teacher@12"
          },
          "image": {
            "type": "string",
            "format": "binary"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "years_of_experience": {
            "type": "number"
          }
        }
      },
...
    }
  }
}